Add 'ghost install' logic and dependencies

+++ install_setup_mysql.sh
@@ -1,5 +1,3 @@
-#!/bin/bash
-
 [ $# > 0 ] && WWW_CODENAME=$1
 WWW_CODENAME=${WWW_CODENAME:-www}
 
@@ -11,16 +9,28 @@ apt-get install -y mysql-server
 mysql_secure_installation
 
 # Read vars
-read -p "Enter MySQL root password [root]: " DB_ROOT_PWD; DB_ROOT_PWD=${DB_ROOT_PWD:-root}
-read -p "Enter MySQL DB username [$WWW_CODENAME]: " DB_USER; DB_USER=${DB_USER:-$WWW_CODENAME}
-read -p "Enter MySQL DB password [$WWW_CODENAME]: " DB_PWD; DB_PWD=${DB_PWD:-$WWW_CODENAME}
-read -p "Enter MySQL DB name [$WWW_CODENAME]: " DB_NAME; DB_NAME=${DB_NAME:-$WWW_CODENAME}
+echo "Set up user and DB?"
+select choice in "Yes" "No"
+do
+	case $choice in
+		Yes ) $UPSTART=true; break;;
+		No ) $UPSTART=false; break;;
+	esac
+done
 
-# Run queries
-mysqlexec='mysql -uroot -p'${DB_ROOT_PWD}' -e'
+if [ $SETUP_USER = 'true' ]; then
 
-$mysqlexec "DROP USER ${DB_USER}@localhost;"
-$mysqlexec "CREATE USER ${DB_USER}@localhost IDENTIFIED BY '${DB_PWD}';"
-$mysqlexec "CREATE DATABASE ${DB_NAME} DEFAULT CHARACTER SET utf8 COLLATE utf8_unicode_ci;"
-$mysqlexec "GRANT ALL ON ${DB_NAME}.* TO ${DB_USER}@localhost;"
-$mysqlexec "FLUSH PRIVILEGES;"
+	read -p "Enter MySQL root password [root]: " DB_ROOT_PWD; DB_ROOT_PWD=${DB_ROOT_PWD:-root}
+	read -p "Enter MySQL WP username [$WWW_CODENAME]: " DB_USER; DB_USER=${DB_USER:-$WWW_CODENAME}
+	read -p "Enter MySQL WP password [$WWW_CODENAME]: " DB_PWD; DB_PWD=${DB_PWD:-$WWW_CODENAME}
+	read -p "Enter MySQL DB name [$WWW_CODENAME]: " DB_NAME; DB_NAME=${DB_NAME:-$WWW_CODENAME}
+
+	# Run queries
+	mysqlexec='mysql -uroot -p'${DB_ROOT_PWD}' -e'
+
+	$mysqlexec "DROP USER ${DB_USER}@localhost;"
+	$mysqlexec "CREATE USER ${DB_USER}@localhost IDENTIFIED BY '${DB_PWD}';"
+	$mysqlexec "CREATE DATABASE ${DB_NAME} DEFAULT CHARACTER SET utf8 COLLATE utf8_unicode_ci;"
+	$mysqlexec "GRANT ALL ON ${DB_NAME}.* TO ${DB_USER}@localhost;"
+	$mysqlexec "FLUSH PRIVILEGES;"
+fi
